feat(ui): add new Home component with advanced cover design UI
- Introduce Home.tsx with 623 lines implementing a comprehensive UI for cover design
- Features include icon search, background color/gradient, cover types, size presets
- Supports overlays: shapes, noise, soft shadow, border, glass panel
- Allows title text customization and preview
- Enables SVG, PNG, and server-rendered image downloads
- Migrated to Next.js App Router and React 18
- Incremental TypeScript conversion for libs/components and main UI
- Updated README with new features and tech upgrades
- Minor fixes in API render route and SVGToImage utility
- Updated next.config.js and tsconfig.json for new packages and TS settings
