+++ README.md
@@ -33,8 +33,13 @@ npm run dev
 - Optional title text overlay (color, size, alignment, position)
 - Shape overlays (blob, circle, rounded stripe) with rotation and opacity
 - Noise grain overlay with intensity and scale
+- Effects: soft-shadow, border (width/radius/color), and glass panel
 - Multi-icon patterns (alternate a secondary icon)
 - Server-side PNG/WebP export for consistent, crisp output
+
+Tech upgrades
+- Migrated to Next.js App Router (`/app`) and React 18
+- Incremental TypeScript conversion (libs/components + main UI)
 
 To export via server, use the “Server PNG” or “WebP” buttons in the UI.
