+++ app/page.tsx
@@ -1,10 +1,6 @@
-"use client";
 import React from "react";
-// Reuse the existing Home component logic for now.
-// The rest of the app (components/libs) is converted to TS; this page will be migrated next.
-import Home from "../pages/index";
+import Home from "./Home";
 
 export default function Page() {
   return <Home />;
 }
-
