+++ next.config.js
@@ -3,4 +3,7 @@ module.exports = {
   images: {
     domains: ["fonts.gstatic.com"],
   },
+  experimental: {
+    serverComponentsExternalPackages: ["@resvg/resvg-js", "sharp"],
+  },
 };
